Fix symfony money form type issue and possible rounding issues

+++ src/Sylius/Bundle/ShippingBundle/Form/Type/Calculator/FlatRateConfigurationType.php
@@ -51,7 +51,7 @@ class FlatRateConfigurationType extends AbstractType
                 'label' => 'sylius.form.shipping_calculator.flat_rate_configuration.amount',
                 'constraints' => array(
                     new NotBlank(),
-                    new Type(array('type' => 'numeric')),
+                    new Type(array('type' => 'integer')),
                 )
             ))
         ;
